[network] Add ValidatorSet to DiscoveryChangeListener
