+++ network/discovery/src/builder.rs
@@ -1,7 +1,7 @@
 // Copyright (c) The Diem Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::ValidatorSetChangeListener;
+use crate::DiscoveryChangeListener;
 use channel::diem_channel;
 use diem_config::network_id::NetworkContext;
 use diem_crypto::x25519::PublicKey;
@@ -13,7 +13,7 @@ use std::sync::Arc;
 use tokio::runtime::Handle;
 
 pub struct ValidatorSetChangeListenerBuilder {
-    listener: Option<ValidatorSetChangeListener>,
+    listener: Option<DiscoveryChangeListener>,
 }
 
 impl ValidatorSetChangeListenerBuilder {
@@ -25,11 +25,11 @@ impl ValidatorSetChangeListenerBuilder {
         reconfig_events: diem_channel::Receiver<(), OnChainConfigPayload>,
     ) -> ValidatorSetChangeListenerBuilder {
         Self {
-            listener: Some(ValidatorSetChangeListener::new(
+            listener: Some(DiscoveryChangeListener::validator_set(
                 network_context,
+                conn_mgr_reqs_tx,
                 expected_pubkey,
                 encryptor,
-                conn_mgr_reqs_tx,
                 reconfig_events,
             )),
         }
@@ -37,7 +37,7 @@ impl ValidatorSetChangeListenerBuilder {
 
     pub fn start(&mut self, executor: &Handle) -> &mut Self {
         let listener = self.listener.take().expect("Listener must be built");
-        executor.spawn(listener.start());
+        listener.start(executor);
         self
     }
 }
